requirements.txt
fastapi
uvicorn
numpy==1.24.3
opencv-python-headless
python-multipart
tensorflow-cpu==2.13.0
Pillow